requirements.txt
Flask
flask-cors
python-twitter